Add a cast

+++ src/partially_blind_rsa.c
@@ -20,7 +20,7 @@
 #include "partially_blind_rsa.h"
 
 #ifndef OPENSSL_IS_BORINGSSL
-#    define BN_bn2bin_padded(OUT, LEN, IN) (BN_bn2binpad((IN), (OUT), (LEN)) == (LEN))
+#    define BN_bn2bin_padded(OUT, LEN, IN) (BN_bn2binpad((IN), (OUT), (LEN)) == (size_t) (LEN))
 #endif
 
 #define MIN_MODULUS_BITS      2048
